build: NetBSD fixes for use of the macro toupper() and tolower()
gcc on NetBSD fails when passing char* to the to{upper,lower} macros.
Cast to a int to work around this.

CR: dalmeida@likewise.com

(centutils: r39716)

+++ centutils/libcentutils/ctstrutils.c
@@ -184,7 +184,7 @@ CTStrToUpper(
 
     if (pszString != NULL) {
         while (*pszString != '\0') {
-            *pszString = toupper(*pszString);
+            *pszString = toupper((int)*pszString);
             pszString++;
         }
     }
@@ -199,7 +199,7 @@ CTStrToLower(
 
     if (pszString != NULL) {
         while (*pszString != '\0') {
-            *pszString = tolower(*pszString);
+            *pszString = tolower((int)*pszString);
             pszString++;
         }
     }
